Extract shared atomic_write for temp-write-then-rename persistence
The write-tmp/rename/fallback-remove-and-rename/cleanup dance was
implemented three times (remote playlist cache persist, cached-bytes
writer, history v2 index). One helper in engine::disk now owns it,
including the Windows rename-over-existing fallback. The playlist
persist path loses only the sub-step "Finalizing cache entry"
progress detail, which flashed for the duration of a rename.
